With this we now stop all docker containers when we stop the servo_ci service.
This is technically a bit too much because we might have other
containers that are not started by the CI but I think this is a valid
change for now.
It also improves the handling of stopping the service. Previously
systemd was waiting for the containers to stop but could not kill them,
so this improves the time needed for servo_ci stop or restart.