This involves a couple of changes:
- Upgrade Dependencies
- Have a new function `check_and_inc_retries` that will panic if the
  retries is too much without success. We will exit the process and
crash. If we succeed we set the retries to zero.
- Unify processor timeout throughout main loop. Some paths through the
  main loop had some timeouts while some had none. We now have a unified
timeout at the end (before killing offline runners) of 5 seconds.
- Read the env variable "RUNNER_SUFFIX" to add a suffix to the runner
  name, so we do not have to hardcode it.
- Introduce a new function `call_github_runner_api` which encapsulates
  simple github api calls with gh. `spawn_runner` now uses this new
functionality for spawning. Previously some
- Introduce `kill_offline_runners` which calsl the github api for
  current runners, takes runners that are offline and have a name
starting with `dresden-hos`. Then it removes these runners. This happens
at the end and 5 seconds after we started the runners. The runners need
some time to start and connect to github properly. The timeout seems ok
tbut we can probably increase it.

Sorry for the massive changes but I think the codebase is so small that
it is allowed :D

Signed-off-by: Narfinger <Narfinger@users.noreply.github.com>