Fix signing config owner
The in container UID/GID are different on our runner machines.
The in-container owner of .ohos was determined by the UID in the tar.gz archive,
which happened to match the in-container servo_ci user on one
of our runners.
We need to manually run `chown` to ensure it always matches,
since apparently `chown` is applied by add before un-tarring the archive.